+++ CHANGELOG.md
@@ -29,6 +29,7 @@ If you like this project and find it useful, please consider giving it a star on
 ### Added
 
 - [service]: Add path of service file.
+- [service]: Add Environment PATH to the service file.
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>